[Chrome] work around restricted 3rd-party storage
If the user has disabled 3rd-party cookies and storage in their Chrome
settings then our iframe can't set cookies and can't read local storage
because we're considered "3rd-party" so long as we're an iframe on a
different origin than our own.

This change asks for the 'contentSettings' permission, which provides
the ability to change these settings programmatically. The extension
now whitelists itself for storage, images, and javascript.